Moves add setups to /setups
